Fix a potential warning about printf format with size_t.

+++ libswfdec-gtk/swfdec_playback_pa.c
@@ -76,7 +76,7 @@ stream_write_callback (pa_stream *pa,
 
   frag = malloc (bytes);
   if (frag == NULL) {
-    g_printerr ("Failed to allocate fragment of size %d\n", bytes);
+    g_printerr ("Failed to allocate fragment of size %d\n", (int)bytes);
     return;
   }
